feat(ids-admin): Various fixes in permissions and client
