+++ libs/portals/admin/ids-admin/src/screens/Client/Client.graphql
