+++ libs/portals/admin/ids-admin/src/screens/Client/Client.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import { EditClient } from '../../components/Client/EditClient'
+import { ClientProvider } from '../../components/Client/ClientContext'
+
+const Client = () => (
+  <ClientProvider>
+    <EditClient />
+  </ClientProvider>
+)
+
+export default Client
