+++ libs/portals/admin/ids-admin/src/screens/Clients/Clients.graphql
