+++ libs/portals/admin/ids-admin/src/screens/Clients/Clients.tsx
@@ -132,7 +132,7 @@ const Clients = () => {
               text={item.defaultEnvironment.clientId}
               tags={item.availableEnvironments.map((tag) => ({
                 children: tag,
-                onClick: () => navigate(href + `?env=${tag}`),
+                onClick: () => navigate(`${href}?env=${tag}`),
               }))}
               eyebrow={<ClientType client={item} />}
               cta={{
