+++ libs/portals/admin/ids-admin/src/shared/hooks/useMultiEnvSupport.tsx
@@ -1,11 +1,9 @@
-import { useContext, useEffect, useState } from 'react'
-import { ClientContext } from '../context/ClientContext'
+import { useEffect, useState } from 'react'
+import { useClient } from '../../components/Client/ClientContext'
 
 export const useMultiEnvSupport = () => {
-  const { client } = useContext(ClientContext)
-  const [shouldSupportMultiEnv, setShouldSupportMultiEnv] = useState<boolean>(
-    true,
-  )
+  const { client } = useClient()
+  const [shouldSupportMultiEnv, setShouldSupportMultiEnv] = useState(true)
 
   useEffect(() => {
     setShouldSupportMultiEnv(client?.availableEnvironments?.length > 1)
