+++ libs/portals/admin/ids-admin/src/shared/hooks/useSyncedQueryStringValueWithoutNavigation.tsx
@@ -5,7 +5,7 @@ import { useEffect, useRef } from 'react'
  * and updating the search params will trigger the search to update for you.
  * However, it also triggers a navigation to the new url, which will trigger
  * the loader to run which we do not want because all our data is already
- * on the client and we're just doing client-side state change on data we
+ * on the client, and we're just doing client-side state change on data we
  * already have. So we manually call `window.history.pushState` to avoid
  * the router from triggering the loader.
  *
